[Silver III] Title: 두 수의 합, Time: 296 ms, Memory: 24904 KB -BaekjoonHub

+++ 백준/Silver/3273. 두 수의 합/두 수의 합.java
@@ -9,11 +9,11 @@ public class Main{
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int n = Integer.parseInt(br.readLine());
         StringTokenizer st = new StringTokenizer(br.readLine());
-        long x = Long.parseLong(br.readLine());
+        int x = Integer.parseInt(br.readLine());
         
-        long[] nums = new long[n];
+        int[] nums = new int[n];
         for(int i=0; i<n; i++) {
-            nums[i] = Long.parseLong(st.nextToken());
+            nums[i] = Integer.parseInt(st.nextToken());
         }
         Arrays.sort(nums);
         
@@ -21,7 +21,7 @@ public class Main{
         int flagE = n-1;
         int count = 0;
         while(flagS < flagE) {
-            long sum = nums[flagS] + nums[flagE];
+            int sum = nums[flagS] + nums[flagE];
             if (x == sum) {
                 count++;
                 flagS++;
